requirements.txt
pyzmq
flask
flask_sqlalchemy
coloredlogs
